perf(exceptions): reduce GC pressure to identify exceptions
This PR introduces caching when processing exceptions.

The objective is to drastically reduce the amount of allocated memory
and allocation operations during exceptions processing.

Two concurrent-safe caches are managed by the exceptions.Processor:
* a regexp cache that saves extraneous regexp.Compile effort
* a cache for AttributesDesignators, which are costly to build

The implementation of the AttributesDesignators cache relies on an
internal hashing method that builds a unique signature for the
attributes (map[string]string). The unitary hash method used is FNV.

The SetExceptions() method of the Result type has been extended to
support (optionally) the reuse of the exceptions.Processor.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>